feat: refactor authz registration
